Fix off-by-one error for last row
Rows get returned starting after the offset, so if the offset is equal
to the number of rows, we're at the end and should return none. The
LimitIterator in combination with SplFileObject will happily let us ask
for offsets past the file end, so this off-by-one actually manifests as
doubling the final row (if the number of "real" rows in the file is a
multiple of the batch size).

(fix #129)