Account for OpenSpout iterator behavior change
The old Spout RowIterator returned null from current() if called before
next() or rewind() (which interally calls next()). We had code for
handling this by calling next() if current() was null, but thanks to
internal changes, the current version is a TypeError.

This commit replaces the null check with a TypeError catch, and uses
rewind() instead of next() so the iterator will correctly refuse a later
rewind.

This rewind change revealed an issue with checkNumberOfColumnsByRow, which
used a foreach after getHeaders(), which calls getRow(). The old code
here allowed the implied foreach rewind(), which would skip a row when
doing the column number check. This commit also includes a fix for that
problem, by switching checkNumberOfColumnsByRow() to use a while loop.